Optimize/ingo (#4)
Optimize main.ts
