+++ __template__/assets/sass/layout/_sidebar.scss
@@ -12,9 +12,9 @@
     position: relative;
 
     &:before {
-      @include vendor("transform", "scaleX(-1)");
+      @include vendor('transform', 'scaleX(-1)');
       color: _palette(fg);
-      content: "\f002";
+      content: '\f002';
       cursor: default;
       display: block;
       font-size: 1.5em;
@@ -28,7 +28,7 @@
       width: calc(_size(element-height) / 1.375);
     }
 
-    input[type="text"] {
+    input[type='text'] {
       padding-right: _size(element-height);
     }
   }
@@ -37,12 +37,9 @@
 #sidebar {
   $pad: calc(2em / 0.9);
 
-  @include vendor("flex-grow", "0");
-  @include vendor("flex-shrink", "0");
-  @include vendor(
-    "transition",
-    ("margin-left 0.5s ease", "box-shadow 0.5s ease")
-  );
+  @include vendor('flex-grow', '0');
+  @include vendor('flex-shrink', '0');
+  @include vendor('transition', ('margin-left 0.5s ease', 'box-shadow 0.5s ease'));
   background-color: _palette(bg-alt);
   font-size: 0.9em;
   position: relative;
@@ -84,7 +81,6 @@
 
   .toggle {
     @include icon(false, solid);
-    @include vendor("transition", "left 0.5s ease");
     -webkit-tap-highlight-color: rgba(255, 255, 255, 0);
     border: 0;
     display: block;
@@ -102,7 +98,7 @@
     z-index: _misc(z-index-base);
 
     &:before {
-      content: "\f0c9";
+      content: '\f0c9';
       font-size: 2rem;
       height: inherit;
       left: 0;
@@ -118,7 +114,7 @@
     margin-left: (_size(sidebar-width) * -1);
   }
 
-  @include breakpoint("<=xlarge") {
+  @include breakpoint('<=xlarge') {
     $pad: calc(1.5em / 0.9);
 
     width: _size(sidebar-width-alt);
@@ -151,7 +147,7 @@
     }
   }
 
-  @include breakpoint("<=large") {
+  @include breakpoint('<=large') {
     box-shadow: 0 0 5em 0 rgba(0, 0, 0, 0.175);
     height: 100%;
     left: 0;
@@ -173,7 +169,7 @@
       top: 0;
 
       &:after {
-        content: "";
+        content: '';
         display: block;
         height: 4em;
         width: 100%;
@@ -195,7 +191,7 @@
     }
   }
 
-  @include breakpoint("<=small") {
+  @include breakpoint('<=small') {
     .toggle {
       text-indent: 7.25em;
       width: 7.25em;
@@ -211,7 +207,7 @@
       &:after {
         background: transparentize(lighten(_palette(fg), 35), 0.25);
         border-radius: _size(border-radius);
-        content: "";
+        content: '';
         height: 3.5em;
         left: 1em;
         position: absolute;
